requirements.txt
absl-py==0.9.0
flask
appnope==0.1.0
astor==0.8.1
attrs==18.2.0
backcall==0.1.0
cachetools==4.0.0
catkin-pkg==0.4.16
certifi==2018.11.29
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.3.2
distro==1.4.0
docutils==0.16
gast==0.3.3
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10
joblib==0.14.1
jsonschema==3.0.0
Kivy==1.11.0
Kivy-Garden==0.1.4
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mglearn==0.1.7
nbformat==4.4.0
numpy==1.16.1
oauthlib==3.1.0
opt-einsum==3.1.0
OWSLib==0.17.1
pandas==0.25.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==3.6.1
prompt-toolkit==3.0.3
protobuf==3.11.1
psycopg2==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pygame==1.9.6
Pygments==2.3.1
pyparsing==2.4.6
PyQt5==5.15.0
PyQt5-sip==12.8.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
pyzmq==19.0.0
QtPy==1.9.0
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
rospkg==1.2.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Shapely==1.6.4.post2
six==1.12.0
termcolor==1.1.0
tornado==6.0.4
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
